backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
sqlmodel>=0.0.8
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
psycopg2-binary>=2.9.7
pydantic-settings>=2.0.0
pytz>=2023.3
pyfiglet>=1.0.2